Skiping tests

+++ src/agents/inference_agent/InferenceRequest.cc
@@ -37,8 +37,8 @@ string InferenceRequest::get_type() { return "INFERENCE_REQUEST"; }
 
 string InferenceRequest::get_max_proof_length() { return to_string(max_proof_length); }
 
-void InferenceRequest::set_timeout(unsigned long long timeout) { this->timeout = timeout; }
-unsigned long long InferenceRequest::get_timeout() { return timeout; }
+void InferenceRequest::set_timeout(unsigned int timeout) { this->timeout = timeout; }
+unsigned int InferenceRequest::get_timeout() { return timeout; }
 
 template <typename T>
 static vector<vector<T>> product(const vector<T>& iterable, size_t repeat) {
